Add files via upload
The program will ask the user to choose from different options based on the circuit configuration.
1. RL circuit.
2. RC circuit.
3. RLC series circuit.
4. RLC parallel circuit.
For RLC circuits the program will choose from three different circuits based on the values of R, L, and C (Over damped, Under damped, or Oscillatory damped).
For each circuit, the user will be asked about the initial conditions (voltage and/or current in the storage element). Based on the circuit configuration the program will solve the problem and find the voltage or the current across the resistor.